requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

biopython==1.79
click==8.1.4
lmdb==1.4.1
msgpack==1.0.7
networkx==3.1
numpy==1.23.5
openbabel-wheel==3.1.1.19
pandas==1.5.3
ProDy==2.4.1
pyrr==0.10.3
python-dotenv==1.0.1
rdkit==2023.3.3
scikit-learn==1.2.2
scipy==1.13.1
torch==2.4.0
torch_geometric==2.3.1
tqdm==4.65.0
wandb==0.15.11
